requirements.txt
# Core Data Science
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
imbalanced-learn==0.11.0
joblib==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Dashboard & Web
streamlit==1.29.0
fastapi==0.105.0
uvicorn[standard]==0.25.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Hugging Face Integration
huggingface-hub==0.19.4
transformers==4.36.0

# Supabase Integration
supabase==2.3.0
postgrest==0.13.0

# Data Processing
openpyxl==3.1.2
python-dotenv==1.0.0

# Jupyter & Notebooks
jupyter==1.0.0
ipykernel==6.27.1
ipywidgets==8.1.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Utilities
python-multipart==0.0.6
requests==2.31.0
tqdm==4.66.1

